Added: Switch component doc.
