+++ content/docs/components/switch.md
@@ -1,0 +1,32 @@
+---
+title: "Switch"
+description: "The Toggle Switch component provides a user-friendly way to enable or disable a setting or feature. It presents a visual representation of two states, typically on and off, allowing users to easily toggle between them. The Toggle Switch component is highly versatile and can be integrated into various user interfaces to enhance user interactions and control options."
+---
+
+## Default switch
+
+The "Default Switch" component provides a visual representation of two states, typically "on" and "off," allowing users to effortlessly toggle between them.
+::DefaultSwitch
+::
+
+## Switch With Icon
+
+Enhance the visual appeal of the switch by adding icons. Simply enable the withIcon with `variant="icon"` prop.
+::SwitchWithIcon
+::
+
+## Disabled Switch
+
+Disable the switch component by utilizing the `disabled` prop with `disabled="true"`.
+::DisabledSwitch
+::
+
+## API Reference
+
+Refer to the API documentation for detailed information about the Toggle Switch component.
+::SwtchApi
+::
+
+## Reference
+
+To learn more about the Switch, please see the documentation of [Redix Vue Switch](https://www.radix-vue.com/components/switch).
