--- pages/docs/components/switch/variant/DisabledSwitch.vue
@@ -1,7 +1,0 @@
-<script lang="ts" setup></script>
-
-<template>
-  <div class="px-5 py-3">
-    <Switch disabled />
-  </div>
-</template>
