--- pages/docs/components/switch/variant/SwitchWithIcon.vue
@@ -1,11 +1,0 @@
-<script lang="ts" setup>
-const switchState = ref(false);
-
-watch(switchState, () => [console.log(switchState.value)]);
-</script>
-
-<template>
-  <div class="px-5 py-3">
-    <Switch v-model:checked="switchState" variant="icon" />
-  </div>
-</template>
